add silo access settings page
