+++ app/layouts/SiloLayout.tsx
@@ -36,7 +36,7 @@ export default function SiloLayout() {
           { value: 'Projects', path: pb.projects() },
           { value: 'Images', path: pb.siloImages() },
           { value: 'Utilization', path: pb.siloUtilization() },
-          { value: 'Silo Access', path: pb.siloAccess() },
+          { value: 'Silo Access', path: pb.siloAccessPolicy() },
         ]
           // filter out the entry for the path we're currently on
           .filter((i) => i.path !== pathname)
@@ -67,7 +67,7 @@ export default function SiloLayout() {
           <NavLinkItem to={pb.siloUtilization()}>
             <Metrics16Icon /> Utilization
           </NavLinkItem>
-          <NavLinkItem to={pb.siloAccess()}>
+          <NavLinkItem to={pb.siloAccessPolicy()}>
             <Access16Icon /> Silo Access
           </NavLinkItem>
         </Sidebar.Nav>
