+++ app/util/date.ts
@@ -7,6 +7,8 @@
  */
 import { formatDistanceToNowStrict, type FormatDistanceToNowStrictOptions } from 'date-fns'
 
+import { pluralize } from './str'
+
 // locale setup and formatDistance function copied from here and modified
 // https://github.com/date-fns/date-fns/blob/56a3856/src/locale/en-US/_lib/formatDistance/index.js
 
@@ -53,3 +55,43 @@ export const toLocaleTimeString = (d: Date, locale?: string) =>
 
 export const toLocaleDateTimeString = (d: Date, locale?: string) =>
   new Intl.DateTimeFormat(locale, { dateStyle: 'medium', timeStyle: 'short' }).format(d)
+
+const SECONDS_MIN = 60
+const SECONDS_HOUR = 60 * 60
+const SECONDS_DAY = 24 * 60 * 60
+
+const numFmt = Intl.NumberFormat()
+
+/**
+ * Formats a duration unit with optional "about" prefix
+ */
+const formatDurationUnit = (
+  seconds: number,
+  unitSeconds: number,
+  unitName: string
+): string => {
+  const units = Math.round(seconds / unitSeconds)
+  const prefix = seconds % unitSeconds === 0 ? '' : 'About '
+  return `${prefix}${numFmt.format(units)} ${pluralize(unitName, units)}`
+}
+
+/**
+ * Formats seconds duration into a human-readable string.
+ * For durations longer than a minute, adds the duration in parentheses.
+ */
+export const formatDurationSeconds = (seconds: number): string => {
+  const secondsStr = `${numFmt.format(seconds)} ${pluralize('second', seconds)}`
+  if (seconds < 60) return secondsStr
+
+  let humanized: string
+
+  if (seconds >= SECONDS_DAY) {
+    humanized = formatDurationUnit(seconds, SECONDS_DAY, 'day')
+  } else if (seconds >= SECONDS_HOUR) {
+    humanized = formatDurationUnit(seconds, SECONDS_HOUR, 'hour')
+  } else {
+    humanized = formatDurationUnit(seconds, SECONDS_MIN, 'minute')
+  }
+
+  return humanized + ` (${secondsStr})`
+}
