+++ app/util/path-builder.spec.ts
@@ -80,7 +80,8 @@ test('path builder', () => {
         "samlIdp": "/system/silos/s/idps/saml/pr",
         "serialConsole": "/projects/p/instances/i/serial-console",
         "silo": "/system/silos/s",
-        "siloAccess": "/access",
+        "siloAccessPolicy": "/access/policy",
+        "siloAccessSettings": "/access/settings",
         "siloIdpsNew": "/system/silos/s/idps-new",
         "siloImageEdit": "/images/im/edit",
         "siloImages": "/images",
